+++ fabric-commands-v1/src/main/java/net/fabricmc/fabric/api/command/v1/DispatcherRegistrationCallback.java
@@ -14,12 +14,23 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.commands;
+package net.fabricmc.fabric.api.command.v1;
 
 import com.mojang.brigadier.CommandDispatcher;
-import net.fabricmc.fabric.api.commands.ServerCommandSource;
 
-public class CommandManagerHolder {
+import net.fabricmc.fabric.api.event.Event;
+import net.fabricmc.fabric.api.event.EventFactory;
 
-	public static CommandDispatcher<ServerCommandSource> COMMAND_DISPATCHER;
+public interface DispatcherRegistrationCallback {
+	Event<DispatcherRegistrationCallback> EVENT = EventFactory.createArrayBacked(
+			DispatcherRegistrationCallback.class,
+			(dispatcher, isDedicatedServer) -> {
+			},
+			callbacks -> (dispatcher, isDedicatedServer) -> {
+				for (DispatcherRegistrationCallback callback : callbacks) {
+					callback.initialize(dispatcher, isDedicatedServer);
+				}
+			});
+
+	void initialize(CommandDispatcher<ServerCommandSource> dispatcher, boolean isDedicatedServer);
 }
